_SsContModulation _SsContPortaTime (#907)
https://decomp.me/scratch/V1zKk
https://decomp.me/scratch/XygVI

+++ src/main/psxsdk/libsnd/ut_sva.c
@@ -1,7 +1,8 @@
 #include "common.h"
 #include "libsnd_internal.h"
 
-short SsUtSetVagAtr(short vabId, short progNum, s32 toneNum, VagAtr* pVagAttr) {
+short SsUtSetVagAtr(
+    short vabId, short progNum, short toneNum, VagAtr* pVagAttr) {
     s16 toneIndex;
     if (svm_vab_used[vabId] == 1) {
         SpuVmVSetUp(vabId, progNum);
